Avoid premature dropping of objects in pdf_dict_put
When inserting a new value into a dictionary, if replacing an existing
entry, ensure we keep the new value before dropping the old one.

This is important in the case where (for example) the existing value
is "[ object ]" and the new value is "object". If we drop the array
and that loses the only reference to object, we can find that we have
lost the value we are adding.