Add routine to clean pdf content streams for pages.
New routine to filter the content streams for pages, xobjects,
type3 charprocs, patterns etc. The filtered streams are guaranteed
to be properly matched with q/Q's, and to not have changed the top
level ctm. Additionally we remove (some) repeated settings of
colors etc. This filtering can be extended to be smarter later.

The idea of this is to both repair after editing, and to leave the
streams in a form that can be easily appended to.

This is preparatory to work on Bates numbering and Watermarking.

Currently the streams produced are uncompressed.