Bug 695171: Pointer error when padding an fz_concat machine
When we return the padding byte in an fz_concat stream, ensure that
we remember to increment rp to point just past in. If not, then we'll
read 2 whitespace chars out. This is fine unless we try and
fz_unread_byte the first one, when we'll leave rp pointing to
an out of buffer address.

Credit to Malc for the bisecting/debugging that got me to the fix.
Many thanks.