Fix 694909: revert "Force colorspaces to match with JPX images." and ...
... instead convert a JPEG2000 used as a soft mask into grayscale.

This is more robust than trusting the PDF specified colorspace over
the internal JPX colorspace.

The spec implies that in a colorspace conflict, the internal JPX
colorspace should be used.

The PDF colorspace may be a DeviceN or Separation colorspace.
DeviceN and Separation colorspaces are not valid destination
colorspaces, so we may not always be able to convert the internal
JPX colorspace into the PDF specified colorspace.

Converting from the internal colorspace into grayscale is more robust,
and solves the issue that the original commit was intended to fix.