Fix crashes, add annotations,  more AA control,  Keyboard commands, select all plus more.
Includes xml, html, text save as, fix issues in window range that determines what
pages are visible at a particular scaling.  Fix very tricky to find problem in
interface with gs.  Managed code was freeing delegates that were allocated dynamically.
It is not possible to pin these so they are now member variables.