Bug #694104 - Allow using MuPDFReaderView from GUI layout tool.
The GUI layout tool instantiates custom controls classes to display a
preview in the IDE. It relies on the 2-argument constructor being
implemented.

Use a different means to get the window manager that works for non-activity
contexts, and avoid creating gesture recognizers in this situation.

Based on a patch supplied by Masaki Muranaka