Fix bug #695346 - iOS app does not register to open CBZ files.
Use the UTI 'public.cbz-archive' instead of the more generic
'public.zip-archive'. It's likely another app (or the OS) has an
Exported UTI for 'public.zip-archive' which would take precedence over
our Imported UTI and not have a 'cbz' extension.

It's not entirely clear what the correct UTI to use for CBZs is (there
are a few), but public.cbz-archive seems to be most likely candidate.

Remove the 'zip' extension, as we probably don't want to load arbitrary
zip files.

Add mimetypes for all completeness, and separate OXPS from XPS as it has
a different mimetype.