hex-encode UTF-16 strings when writing PDF
fmt_obj calculates whether a string is better hex-encoded or written
using escapes. Due to a bug, '\0' is considered to be escapable same as
'\n' when instead it would have to be written as '\000'. Since UTF-16
strings tend to consist of many '\0' bytes, their octal encoded form is
much longer than their hex encoded form.

The issue is that the first argument to strchr contains an unintended
trailing '\0' which has to be special-cased first.