Revise test-device; thresholding and exhaustive checking.
The original version of the test-device could characterise pages
as being grayscale/color purely based on the colorspaces used.

This could easily be upset by grayscale images or shadings that
happened to be specified in non-grayscale colorspaces however.

We now look at the actual shading and image color values, and use
a threshold value to allow for some measure of rounding errors in
color values that are in practice grayscale.