Performance improvement for test-device.
When we detect that a page is color, set the ignore image hint
to avoid us loading future images. The overhead on loading
images is not generally huge, except for JPEG2000 ones, which
currently require decoding at load time. This therefore saves
lots of time for such files.

Also, a tiny tweak to ignore page components with 0 alpha.