Update test-device to check images piecemeal.
Rather than decoding entire image only to give up after we find the
very first pixel is color, add code so that the test-device can
treat the image as a stream. This means that (for most image types
at least) we can bale out without decoding everything.

This reduces the runtime of 3001Pages.pdf from 14 minutes to 18 seconds.