Disable options that are not currently possible to do when the source is an image.
Currently if the source was a png, jpeg or cbz file we can't do conversions with
ghostscript and also we can't print.  So I will disable the controls for now and
we can look into adding that functionality later if we want.

Disable printing capability for image based formats.

Currently printing is all XPS based.  If the source
document is JPG, PNG or CBZ we don't currently have
a method for creating the XPS content for these.  We
could do a bitmap print or possibly wrap up the images
in a XPS file or just do a direct bitmap print via some
C# magic.  I need to investigate.  For the short time,
though it is simply disabled.