Rejig display device to have a more compact memory representation.
Rather than a linked list of display nodes, we use solid block of
serialised data. We send a 32bit word, which contains various bitfields.
These bitfields indicate the command type, and the presence or absence
of various fields (such as paths, colorspaces, colors etc).

If these fields are not present, they are held to be the same as the
previous values.