Bug 695843: Tweak bboxes of type3 fonts; honour the d1 values.
The example file for this bug has an invalid font bbox. The current
code uses this bbox (or some multiple of it) to clip the glyphs
size.

In the new code, when we convert the glyphs to display lists we
watch for the bbox given in any d1 operator used. If we find one,
we gather the rectangle specified and store it as the glyph rectangle
in the fz_font.

If we then attempt to bound a glyph that used d1, it happens instantly
without needing to run the list. This seems to match acrobats behaviour.

Tests indicate that Acrobat never clips d0 glyphs, so our behaviour
is still different here, but I am not changing this at the moment.

Also, I note that t3flags should be a un unsigned short but are currently
just a char. Fix that too.

Also fix some missing code in fz_new_font that would cause leaks if mallocs
failed.