Update display list tile skipping.
When we meet cached tiles when rendering the display list, we need to
skip over their contents. Previously we did this by skipping
display list nodes in their entirety.

With the new display list scheme however, we cannot simply skip
nodes completely as the graphic state changes must be remembered.

We therefore update the list playback routine to keep track of the
clip depth and to skip the function calls as required.