Bug 695853: Fix pdf clean operation with invalid refs in input file.
MuPDF (and other PDF readers) treat invalid references as 'null'
objects. For instance, in the supplied file, object 239 is supposedly
free, but a reference is made to it.

When cleaning (or linearising) a file, we renumber objects; such
illegal refs then end up pointing somewhere else.

The workaround here is simply to spot the invalid refs during the
mark phase, and to set the referencing to null.