Reduce pdf_obj memory usage.
Historically pdf_obj was a structure with a header and a union in it.
As time has gone by more stuff has been put into the header, and the
different arms of the union have changed in size. We've even adopted
the idea of different 'kinds' of pdf_obj's being different sizes
(names and strings for examples).

Here we rework the system slightly; we minimise the header, and split
out everything into different structures. Every different 'kind' of
pdf_obj is now it's own structure, just as big as it needs to be.

Key changes:

 * refs is now a short rather than an int. We are never going
   to need more than 32767 refs (indeed, if we ever need more
   than about 3 (10 at the outside), something has gone very
   wrong!). This aids structure packing.

 * Only arrays, dicts and refs actually need the pdf_document
   pointer.

 * Only arrays and dicts need the parent_num pointer.