Bug 694367: Attempt to avoid dropouts of rectangles.
This is not a complete general fix for features dropping out of
rendered line art, but merely a fix for one of the more common
cases.

When rendering rectangles (currently, specifically only those
rectangles that are actually defined as rectangles within the
path structure), if they are axis aligned, then ensure that they
always fill the subpixel line they are on.