Add some basic page margins to EPUB layout.
Add margins of 1 em at the top and bottom of every page at the top level.

TODO: This should be set from the CSS using the @page selector, and be
collapsed with the body margins at the start and end of each chapter, as well
as the left and right body margins.