Fix bug in pdf_dict_find.
Sebras and Tor spotted that we could get occasional 'warning: cannot seek
backwards' messages. An example command that shows this is:

   mutool show pdf_reference17.pdf grep

They further tracked the problem down to the 'sorted' side of the
pdf_dict_find function.

In the binary search, I calculate c to be the comparison value between
pairs of keys. In the case where both keys (names) are in the special
case 'known' range below PDF_OBJ__LIMIT, I use pointer arithmetic for
this. Unfortunately, I was forgetting that the compiler thinks that
pdf_obj *'s are 4 (or 8) bytes in size, so was doing (a-b)/4.

To workaround this I cast both keys to char *'s. This solves the bug.

Thanks to Sebras and Tor for doing the hard work in tracking this down.