pdf_dict_find optimisation.
When doing pdf_dict_put, we first call pdf_dict_find to hunt for an
existing entry we can just update.

Recently we introduced a 'location' return from pdf_dict_find that
would (in the non-found case) return the location of where such an
entry should be inserted.

It's just dawned on me that we don't need a separate variable for
this. We continue to return negative numbers for 'not found', but
these negative numbers can contain the insertion point.