Bug 696053: Update windows mupdf to respect command line flags.
Previously, only the unix executable had been updated to take
command line flags; update the windows one in line with it.

We have to cope with the argv being in Unicode; add a windows
specific version of getoptw for this.

Also note that that fprintf's in the windows mupdf exe won't work
as GUI apps don't have a console window, and can't write to the
parent one. Fixing that is a larger project than I have time
for right now.