Further tweaks to fz_image handling.
Ensure that subsampling and caching happen in the generic image
code, not in the specific.

Previously, the subsampling happened only for images that were
decoded from streams. Images that were loaded direct were never
subsampled and hence were always cached at full size. After this
change both classes of image are correctly subsampled, and
the subsampled version kept in the cache.

This produces various image diffs in the cluster, none of which
are noticable to the naked eye.