GProof: Use better RGB -> CMYK conversion when working from seps.
When proofing an image without all the separations enabled, we
reconstitute it using the equivalent cmyk colors, and then
convert to rgb. The CMYK -> RGB step is important for quality,
and the whole point of GProof mode is to give the best possible
quality, so use the slower, but prettier code.

Also, fix the blatent bugs in the slower but prettier code that
I introduced when hurriedly fixpointing it.