Add stdout and stderr procedures for use with gs api calls
gsview on windows does not support the use of stdout and stderr.
The calls to write out to these will return 0 leading to an i/o
error from gs.   To avoid this we define dummy functions that
simply return the length that was to be written.  To avoid
issues with including "iapi.h" in Android builds that do not
support gproof we wrap all of gprf-doc.c in an ifdef SUPPORT_GPROOF