Support several levels of incremental xref
This fixes bug #696123 by allowing multiple signatures each to be written
to the document in a separate incemental update.

Add count num_incremental_sections to keep track of the number of
incremental sections.

Add xref_base, which can be set between 0 and num_incremental_sections
inclusive to access different versions of the document.

Add disallow_new_increments flag that stops new incremental sections
being provoked by the creation of an xref stream.

Move the unsaved_sigs list from the document structure to the xref
structure. With this commit in place, the lists will never grow beyond
length one, but we've maintained the list structure in case other cases
need supporting in the future.

Add an end offset field to the xref structure, so that during completion
of signatures the document length of the various incremental versions of
the document are available.

Factor out functions for storing unsaved signatures and for checking if
an object is an unsaved signature.

Do deep copy of objects that require the holding of several versions.