Bug 696115: Fix problem with fz_try/fz_catch with modern gcc's.
Modern gcc's have a compiler (optimiser) bug that can cause
values not to be written back to memory when they should. We
work around this by using an inline function to force the
compiler to behave.

Many thanks to Marcos Woehrmann for doing the analysis that
lead to this workaround.