Bug 696129: Be more forgiving of broken files.
The PDF spec says that old format xrefs should start with:

  xref\n<start> <len>

The example file in question has:

  xref <start> <len>

which confuses our parsing code. Update the parse code to avoid
using fz_read_line, and to instead work on a char level.

Also, downgrade the error given when the first object is not free
to be a warning. Now we do 'just in time' repair, we are probably
better able to cope.