Fix incremental xref stream.
When writing incremental xref streams, the opts->use_list entry of the last
object, i.e. of the xref stream object itself, was left uninitialized. This
resulted in a random value, 0 or 1, being written into the xref stream.

Also, always write a newline before the endstream keyword, as that shall be
done for xref streams and should be done for all other streams.