Add mutool create tool, and PDF font and image resource creation.
Initial framework for creating pdfs

This adds a create option to mutool for us to use in working
on the API for creating content as well as adding content to
existing documents.

mutool create: Get page sizes and add them

Start the parsing of the contents.txt file which may have
multiple page information.  Add the pages at the proper sizes.

Further work on mutool create_pdf

Remove the calls that were being made to the pdf-write device.
Clean up several issues with the reading of the page contents.
Get the content streams for each page associated with the page->contents
Temp. created a pdf_create_page_contents procedure.  I will merge
this with pdf_create_page as there is significant overlap.

Next is to add in the font and image resources and indirect references.

Include pdfcreate in build

Merge pdf_create_page_contents and pdf_create_page

Add support for images in pdfcreate

This adds images to the pdf document using a function stolen from pdf-device (send_image).
This was renamed pdf_add_image_res and added to pdf-image.   Down the road, send-image will
be removed.  Prior to that, I need to work on making sure that multiple copies of the same
image do not end up in the document.

Code was also added to create the page resources to point to the proper image in the document.

Next fonts will be added in a similar manner, then I will work on computing the md5 sums of
image and fonts to ensure only one copy ends up in the document.  Then pdf-write will be
reworked to use the same code as opposed to its current list of md5 sums that are stored in
a device structure.

mutool pdfcreate: support for WinAnsiEncoded fonts

Added support for very simple fonts (WinAnsiEncoding).  Methods
added in pdf-font.c.  Added first_width and last_width to fz_font_s
and stem_v to pdf_font_desc_s.

Ran code through memento with simple test of 4 page document
creation including an image and a font.  Fixed several leaks
as well as buffer corruption issues (main changes in pdfcreate).
Thanks to Robin for the help with Memento in finding leaks.

Added StemV to pdf names as it was needed for the font descriptor creation.

Fix for pdf_write_document rename to pdf_save_document

Add resource_ids to pdf document structure

The purpose of this structure will be to allow the search
and reuse of resources when we attempt to add new ones
to the document.

Fix name changes from recent updates

pdf_create branch updated to work with recent changes in master

Initial use of hash table for resources

To avoid adding in the same resource this adds a
resource_tables member to pdf_document.  The
resource_tables structure consists of multiple
fz_hash_table entries, one for each resource type.
When an attempt is made to search for an existing
resource, the table will be initialized in a brute
force search for existing resources.  Currently this
is only set up for the image resources and accessed
through pdf_add_image_res.  If a match is found,
the reference object is returned.  If no match is found
NULL is returned and the ref object created in pdf_add_image_res
is added into the hash table.  In this case, a command line
such as
create -o output.pdf -f F0:font.ttf -i Im0:image.jpg -i Im1:image1.jpg \\
-i Im2:image.jpg contents.txt
will avoid the insertion of two copies of image.jpg into the
output PDF document.

CID Identity-H Font added for handing ttf

This adds a method for adding a ttf to a PDF as a
CID font with Identity-H mapping and a ToUnicode
entry that is created using FT_Get_Char_Index

This takes much care in the creation of the ToUnicode
CMap to ensure that the minimum number of entries
are created in that we try to use beginbfrange as
much as possible before using beginbfchar.  The
code makes sure to limit the number of entries in
a group to 100 and to not cross first-byte boundaries
for the CID values as described in the Adobe
Technical note 5411.

Add missing file pdf-resources.c

pdf-resources.c was missing and should have been
committed earlier.  Added to windows project file.
Not sure where else it needs to be added for the
other platforms.

Clean up names and spacing

Make sure that the visible functions have the proper namespace (e.g. pdf_xxxx)
Also make sure we have a blank line prior to comment.
Be consistent with static function naming in pdf_resources.c

pdfwrite make use of image resource fz_hash_table

The pdfwrite device now shares the structure that stores the
resource images for pdfcreate.   With this fix, pdfwrite now
avoids duplicating the writing of the same images that are
shared across multiple pages.

Add missing file pdf-resources.c

Initial work toward having pdfwrite use Identity-H Type0 encoding for fonts

Finish of CID type0 Identity-H font for pdfwrite

This adds in the proper widths which may have been stored in the source font
in the width table (parsed from the W entry in the pdf file) or if the
free type structure has its own cmap then we can get the width from free type.
Widths are restructured into format described in 5.6.3 of PDF spec.

Fix issue from conflict merging and multiple define of structure

Clean up warnings and make mutool create use simple font